image_gen/requirements.txt
google-genai>=1.0.0
Pillow
python-dotenv
